requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
arrow==1.3.0
attrs==23.2.0
bravado==10.4.1
bravado-core==6.1.1
Bybit==0.2.12
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
dateparser==1.2.0
fqdn==1.5.1
frozenlist==1.4.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httplib2==0.22.0
idna==3.7
importlib_resources==6.4.0
isoduration==20.11.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
monotonic==1.6
msgpack==1.0.8
msgpack-python==0.5.6
multidict==6.0.5
mysql-connector-python==9.0.0
oauth2client==4.1.3
oauthlib==3.2.2
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.0
pybit==5.8.0
pycryptodome==3.20.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
rsa==4.9
setuptools==70.3.0
simplejson==3.19.2
six==1.16.0
swagger-spec-validator==3.0.4
Telethon==1.36.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3>=1.24.2
webcolors==24.6.0
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
yarl==1.9.4
